[HW-1] rename files
