+++ homework_1.py
