+++ homework_2.py
